Updates to use nested slots for the body. Updates naming convention to match other components.
