+++ app/components/buildout_design_system/card.html.slim
@@ -5,8 +5,6 @@
     = header
   = content
   - if body?
-    - if title?
-        = title
     = body
   - if footer?
     = footer
