+++ app/components/buildout_design_system/card.rb
@@ -1,10 +1,9 @@
 module BuildoutDesignSystem
   class Card < ViewComponent::Base
-    renders_one :header, "CardHeaderComponent"
-    renders_one :image, "CardImageComponent"
-    renders_one :title, "CardTitleComponent"
-    renders_one :footer, "CardFooterComponent"
-    renders_one :body, "CardBodyComponent"
+    renders_one :header, "CardHeader"
+    renders_one :image, "CardImage"
+    renders_one :footer, "CardFooter"
+    renders_one :body, CardBody
 
     def initialize(class_name: "", **attrs)
       super(**attrs)
@@ -12,7 +11,7 @@ module BuildoutDesignSystem
       @attrs = attrs
     end
 
-    class CardHeaderComponent < ViewComponent::Base
+    class CardHeader < ViewComponent::Base
       attr_reader :class_name
 
       def initialize(class_name: "", **attrs)
@@ -28,7 +27,7 @@ module BuildoutDesignSystem
       end
     end
 
-    class CardImageComponent < ViewComponent::Base
+    class CardImage < ViewComponent::Base
       attr_reader :image_url, :class_name
 
       def initialize(image_url: "", class_name: "")
@@ -42,22 +41,7 @@ module BuildoutDesignSystem
       end
     end
 
-    class CardTitleComponent < ViewComponent::Base
-      attr_reader :class_name
-
-      def initialize(class_name: "", title: "", **attrs)
-        super()
-        @class_name = class_name
-        @title = title
-        @attrs = attrs
-      end
-
-      def call
-        content_tag :h5, @title, { class: "card-title #{class_name}", **@attrs}
-      end
-    end
-
-    class CardFooterComponent < ViewComponent::Base
+    class CardFooter < ViewComponent::Base
       attr_reader :class_name
 
       def initialize(class_name: "", **attrs)
@@ -68,20 +52,6 @@ module BuildoutDesignSystem
 
       def call
         content_tag :div, content, { class: "card-footer d-flex gap-2 #{class_name}", **@attrs }
-      end
-    end
-
-    class CardBodyComponent < ViewComponent::Base
-      attr_reader :class_name
-
-      def initialize(class_name: "", **attrs)
-        super()
-        @class_name = class_name
-        @attrs = attrs
-      end
-
-      def call
-        content_tag :div, content, { class: "card-body #{class_name}", **@attrs }
       end
     end
   end
