+++ app/components/buildout_design_system/card/card_body.html.slim
@@ -1,0 +1,4 @@
+.card-body class="#{@class_name}" *@attrs
+  - if title?
+    = title
+  = content
