+++ test/dummy/test/components/previews/buildout_design_system/card_preview/with_image.html.slim
@@ -1,9 +1,10 @@
 = render(BuildoutDesignSystem::Card.new()) do  |card|
   - card.with_image(image_url: image_url)
 
-  - card.with_title(title: title, class_name: "p-4")
+  / - card.with_title(title: title, class_name: "p-4")
 
-  - card.with_body() do
+  - card.with_body do | card_body |
+    - card_body.with_title(title: "Card Title")
     = body
 
   - card.with_footer(id: "fancyFooterID") do
